requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
openai
chromadb
pymupdf
tiktoken
python-multipart
pytest
httpx
